file upload hot fix

+++ api/services/UploadService.js
@@ -17,9 +17,9 @@ const pluralize = require('pluralize');
 var slug = require('slug');
 
 module.exports = {
-    createFile: function (req, res, fileRequired, cb) {
+    createFile: function(req, res, fileRequired, cb) {
 
-        var uploadedFile = req.file('uploadFile').on('error', function (err) {
+        var uploadedFile = req.file('uploadFile').on('error', function(err) {
             if (!res.headersSent) return res.negotiate(err);
         });
 
@@ -33,20 +33,20 @@ module.exports = {
         }
     },
 
-    uploadFile: function (req, res, uploadedFile, fileRequired, cb) {
+    uploadFile: function(req, res, uploadedFile, fileRequired, cb) {
         var mimetype = '';
         var extension = '';
         var dataset = req.param('dataset');
         var data = actionUtil.parseValues(req);
         var allowedTypes;
         if (uploadedFile.isNoop) {
-            cb(data);
+            return cb(data);
         } else {
-            Dataset.findOne(dataset).then(function (dataset) {
+            Dataset.findOne(dataset).then(function(dataset) {
 
-                FileType.find().exec(function (err, filetypes) {
+                FileType.find().exec(function(err, filetypes) {
                     if (err) return res.negotiate(err);
-                    allowedTypes = _.transform(filetypes, function (allowedTypes, filetype) {
+                    allowedTypes = _.transform(filetypes, function(allowedTypes, filetype) {
                         allowedTypes.push(filetype.mimetype);
                     }, []);
 
@@ -54,11 +54,13 @@ module.exports = {
                         lower: true
                     });
                     uploadedFile.upload({
-                            saveAs: function (file, cb) {
+                            saveAs: function(file, cb) {
 
                                 //Get the mime and the extension of the file
                                 mimetype = mime.lookup(file.filename.split('.').pop());
                                 extension = file.filename.split('.').pop();
+                                data.fileName += '.' + extension;
+
                                 // If the mime is present on the array of allowed types we can save it
                                 if (allowedTypes.indexOf(mimetype) === -1) {
                                     return res.negotiate({
@@ -71,19 +73,24 @@ module.exports = {
                                     //If file exists, deleted it
                                     if (!fileRequired) {
                                         const pk = actionUtil.requirePk(req);
-                                        File.findOne(pk).populate('dataset').then(function (file) {
-                                            var upath = path.resolve(sails.config.odin.uploadFolder +
-                                                '/' + slug(file.dataset.name, {
-                                                    lower: true
-                                                }) + '/' + file.fileName);
-                                            fs.lstat(upath, function (err, stats) {
-                                                if (!err && stats.isFile()) {
-                                                    UploadService.deleteFile(file.dataset.id, file.fileName, res);
-                                                }
-                                            });
+                                        File.findOne(pk).populate('dataset').then(function(file) {
+                                            // if the uploaded name is the same of the one saved on the filesystem
+                                            // don't deleted, just overwrite it
+                                            console.dir(file.fileName);
+                                            console.dir(data.fileName);
+                                            if (file.fileName !== data.fileName) {
+                                                var upath = path.resolve(sails.config.odin.uploadFolder +
+                                                    '/' + slug(file.dataset.name, {
+                                                        lower: true
+                                                    }) + '/' + file.fileName);
+                                                fs.lstat(upath, function(err, stats) {
+                                                    if (!err && stats.isFile()) {
+                                                        UploadService.deleteFile(file.dataset.id, file.fileName, res);
+                                                    }
+                                                });
+                                            }
                                         });
                                     }
-                                    data.fileName += '.' + extension;
                                     return cb(null, data.fileName);
                                 }
                             },
@@ -106,7 +113,7 @@ module.exports = {
                             // Get the id of the filetype based on mime of the file
                             sails.models.filetype.findOne({
                                 name: extension
-                            }).exec(function (err, record) {
+                            }).exec(function(err, record) {
                                 if (err) return res.negotiate(err);
                                 if (!record) {
                                     return res.serverError('Could not find the filetype uploaded: ' + extension);
@@ -116,59 +123,59 @@ module.exports = {
                                 // If the file is consumable via the API
                                 if (record.api) {
                                     var filePath = sails.config.odin.uploadFolder + "/" + slug(dataset.name, {
-                                            lower: true
-                                        }) + '/' + data.fileName;
+                                        lower: true
+                                    }) + '/' + data.fileName;
 
                                     // Read the file
                                     fs.createReadStream(filePath)
                                         // Encode it
-                                        .pipe(iconv.decodeStream(sails.config.odin.defaultEncoding)).collect(function (err, result) {
-                                        if (err) return res.negotiate(err);
+                                        .pipe(iconv.decodeStream(sails.config.odin.defaultEncoding)).collect(function(err, result) {
+                                            if (err) return res.negotiate(err);
 
-                                        if (sails.config.odin.defaultEncoding === 'utf8') result = '\ufeff' + result;
+                                            if (sails.config.odin.defaultEncoding === 'utf8') result = '\ufeff' + result;
 
-                                        //Should check which type the file is and convert it .
-                                        var json = [];
-                                        if (extension === 'xls' || extension === 'xlsx') {
-                                            //Convert XLS to json, store on nosql database
-                                            var workbook = XLSX.readFile(files[0].fd);
+                                            //Should check which type the file is and convert it .
+                                            var json = [];
+                                            if (extension === 'xls' || extension === 'xlsx') {
+                                                //Convert XLS to json, store on nosql database
+                                                var workbook = XLSX.readFile(files[0].fd);
 
-                                            //Join all the worksheets on one json
-                                            json = _.reduce(workbook.SheetNames, function (result, sheetName) {
-                                                var worksheet = workbook.Sheets[sheetName];
+                                                //Join all the worksheets on one json
+                                                json = _.reduce(workbook.SheetNames, function(result, sheetName) {
+                                                    var worksheet = workbook.Sheets[sheetName];
 
-                                                var currentJson = XLSX.utils.sheet_to_json(worksheet);
-                                                result = _.concat(result, currentJson);
-                                                return result;
-                                            }, []);
-                                            DataStorageService.mongoSave(dataset.id, data.fileName, json, res);
+                                                    var currentJson = XLSX.utils.sheet_to_json(worksheet);
+                                                    result = _.concat(result, currentJson);
+                                                    return result;
+                                                }, []);
+                                                DataStorageService.mongoSave(dataset.id, data.fileName, json, res);
 
 
-                                        } else {
-                                            // Convert to JSON
+                                            } else {
+                                                // Convert to JSON
 
-                                            var converter = new Converter({
-                                                delimiter: 'auto'
-                                            });
-
-                                            converter.fromString(result, function (err, json) {
-                                                if (err) {
-                                                    return res.negotiate(err);
-                                                }
-                                                if (json.length === 0) return res.badRequest(null, {
-                                                    message: "Invalid or empty csv."
+                                                var converter = new Converter({
+                                                    delimiter: 'auto'
                                                 });
 
-                                                // Connect to the db
-                                                DataStorageService.mongoSave(dataset.id, data.fileName, json, res);
-                                                //If file is required the method was update,
-                                                // then we update their visualizations
-                                                if (!fileRequired) {
-                                                    VisualizationsUpdateService.update(data)
-                                                }
-                                            });
-                                        }
-                                    });
+                                                converter.fromString(result, function(err, json) {
+                                                    if (err) {
+                                                        return res.negotiate(err);
+                                                    }
+                                                    if (json.length === 0) return res.badRequest(null, {
+                                                        message: "Invalid or empty csv."
+                                                    });
+
+                                                    // Connect to the db
+                                                    DataStorageService.mongoSave(dataset.id, data.fileName, json, res);
+                                                    //If file is required the method was update,
+                                                    // then we update their visualizations
+                                                    if (!fileRequired) {
+                                                        VisualizationsUpdateService.update(data)
+                                                    }
+                                                });
+                                            }
+                                        });
                                 }
                                 cb(data);
 
@@ -181,11 +188,11 @@ module.exports = {
     },
 
 
-    uploadImage: function (req, res, cb) {
+    uploadImage: function(req, res, cb) {
         var data = actionUtil.parseValues(req);
         var path = sails.config.odin.uploadFolder + '/categories';
 
-        var uploadFile = req.file('uploadImage').on('error', function (err) {
+        var uploadFile = req.file('uploadImage').on('error', function(err) {
             if (!res.headersSent) return res.negotiate(err);
         });
         if (!uploadFile.isNoop) {
@@ -194,7 +201,7 @@ module.exports = {
             });
 
             uploadFile.upload({
-                saveAs: function (file, cb) {
+                saveAs: function(file, cb) {
                     var mimetype = mime.lookup(file.filename.split('.').pop());
 
                     if (mimetype !== 'image/svg+xml') {
@@ -223,7 +230,7 @@ module.exports = {
         }
     },
 
-    metadataSave: function (model, data, modelName, req, res, extraRecordsResponse) {
+    metadataSave: function(model, data, modelName, req, res, extraRecordsResponse) {
         model.create(data).exec(function created(err, newInstance) {
             if (err) return res.negotiate(err);
 
@@ -243,7 +250,7 @@ module.exports = {
 
                 // Make sure data is JSON-serializable before publishing
                 var publishData = _.isArray(newInstance) ?
-                    _.map(newInstance, function (instance) {
+                    _.map(newInstance, function(instance) {
                         return instance.toJSON();
                     }) :
                     newInstance.toJSON();
@@ -252,13 +259,13 @@ module.exports = {
 
             var associations = [];
 
-            _.forEach(model.definition, function (value, key) {
+            _.forEach(model.definition, function(value, key) {
                 if (value.foreignKey) {
                     associations.push(key);
                 }
             });
 
-            model.find(newInstance.id).populate(associations).exec(function (err, record) {
+            model.find(newInstance.id).populate(associations).exec(function(err, record) {
                 if (!_.isUndefined(extraRecordsResponse)) {
                     record[0] = _.merge(record[0], extraRecordsResponse);
                 }
@@ -279,7 +286,7 @@ module.exports = {
 
     },
 
-    metadataUpdate: function (model, data, modelName, req, res, extraRecordsResponse) {
+    metadataUpdate: function(model, data, modelName, req, res, extraRecordsResponse) {
         // Look up the model
         model.update(data.id, data).exec(function updated(err, records) {
 
@@ -320,13 +327,13 @@ module.exports = {
 
             var associations = [];
 
-            _.forEach(model.definition, function (value, key) {
+            _.forEach(model.definition, function(value, key) {
                 if (value.foreignKey) {
                     associations.push(key);
                 }
             });
             //populate the response
-            model.find(updatedRecord.id).populate(associations).exec(function (err, record) {
+            model.find(updatedRecord.id).populate(associations).exec(function(err, record) {
                 if (err) return res.negotiate(err);
 
                 //if we have any extraRecords to add to the response,
@@ -351,14 +358,14 @@ module.exports = {
 
     },
 
-    deleteFile: function (datasetId, fileName, res) {
-        Dataset.findOne(datasetId).then(function (dataset) {
+    deleteFile: function(datasetId, fileName, res) {
+        Dataset.findOne(datasetId).then(function(dataset) {
 
             var path = sails.config.odin.uploadFolder + '/' + slug(dataset.name, {
-                    lower: true
-                }) + '/' + fileName;
+                lower: true
+            }) + '/' + fileName;
 
-            fs.unlink(path, function () {
+            fs.unlink(path, function() {
                 DataStorageService.deleteCollection(dataset.id, fileName, res);
                 ZipService.createZip(dataset.id);
             });
